Add test_computational_basis_state.py
